\name{diet-internal}
\alias{explore}
\alias{explore.bag}
\alias{mask}
\alias{na.dpart}
\alias{na.igam}
\alias{outside}
\alias{plotG.dpart}
\alias{readpred}
\alias{rpartco.dpart}
\alias{rsq}
\alias{rsq.dpart}
\alias{select.tree}
\alias{select.tree.dpart}
\alias{snip.dpart.mouse}
\alias{snip.dpart}
\alias{tree.depth.dpart}
\alias{tree.depth}
\alias{text.dpart}
\alias{textG.dpart}
\alias{vis.igam}
\alias{vis}
\alias{writepn.csv}
\alias{formOmat}

\title{Internal functions used in the diet package}
\description{
These are internal functions used in the \code{diet} packagethat are not intended to be directly called by the user
}


\author{Petra Kuhnert}
\keyword{internal}
