% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorPlots.r
\name{FactorPlots}
\alias{FactorPlots}
\title{FactorPlots}
\usage{
FactorPlots(x, XvarF, i, PredSpID)
}
\arguments{
\item{x}{x}

\item{XvarF}{XvarF}

\item{i}{i}

\item{PredSpID}{PredSpID}
}
\description{
Internal function used by \code{plot.diet}
}
