% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.r
\name{grab}
\alias{grab}
\alias{grabmulti.dpart}
\alias{grabmulti}
\alias{grab.dpart}
\alias{grab.bag}
\title{Summarises a Node of the Tree}
\usage{
grab(object, LatID, LonID, setID = NULL, node.cols = NULL, cex = 0.8,
  mapxlim = NULL, mapylim = NULL, mapcol = "gold3",
  database = "world", onepage = TRUE, display.object, oob = FALSE,
  ylim)

\method{grabmulti}{dpart}(object, LatID, LonID, setID = NULL,
  node.cols = NULL, cex = 0.8, mapxlim = NULL, mapylim = NULL,
  mapcol = "black", n = nrow(object$frame), database = "world", ...)

grabmulti(object, LatID, LonID, setID = NULL, node.cols = NULL,
  cex = 0.8, mapxlim = NULL, mapylim = NULL, mapcol = "gold3",
  n = nrow(object$frame), database = "world", ...)

\method{grab}{dpart}(object, LatID, LonID, setID = NULL,
  node.cols = NULL, cex = 0.8, mapxlim = NULL, mapylim = NULL,
  mapcol = "black", database = "world")

\method{grab}{bag}(object, LatID, LonID, setID = NULL,
  node.cols = NULL, cex = 0.8, mapxlim = NULL, mapylim = NULL,
  mapcol = "gold3", database = "world", display.object, oob = FALSE,
  ylim)
}
\arguments{
\item{object}{tree object either of class \code{dpart} or \code{bag}}

\item{LatID}{string. Column name of latitude.}

\item{LonID}{string. Column name of longitude.}

\item{setID}{optional set identification number.}

\item{node.cols}{vector of node colours assigned to each prey. 
Use output from \code{apc} function. If node colours are not provided, 
the trees will not be produced in colour.}

\item{cex}{numeric. Size of plotting symbols and labels (default: 0.8)}

\item{mapxlim}{optional map x-limits. If none are specified then the 
range of the data are used.}

\item{mapylim}{optional map y-limits. If none are specified then the 
range of the data are used.}

\item{mapcol}{map land colour. (default: gold3)}

\item{database}{either 'world' or 'world2' are available for plotting. 
Defaults to 'world' if
not specified.}

\item{display.object}{tree object to display. This may be different 
to \code{object} when bagging is used.}

\item{oob}{logical. Option when a bagged tree is passed to the 
\code{grab.bag} function to 
use out of bag estimates.}

\item{ylim}{y-axis limit for barcharts that are produced by the 
\code{explore} function}

\item{n}{numeric. Maximum number of nodes to investigate in the tree. 
Defaults to the number in the tree.}
}
\value{
Summary output from the node/s selected in the tree 
consisting of the node number,
number of observations, number of sets (if the setID was provided), number of
predators, number of prey, deviance, expected loss and predicted class.
}
\description{
The \code{grab} functions allow you to select nodes of the tree for further 
investigation and summary. Once a node is selected, a map of the points is shown
with highlighted points of the selection along with a barplot of the prey
distribution found at that node. When applied to an object of class \code{bag},
bagged distributions of prey are shown. The \code{grabmulti} function produces
multiple summaries on successive nodes selected by the user.
}
\details{
The \code{grab.dpart} function can be used on any 
tree object of class \code{dpart}.
  The \code{grab.bag} function can be used on any tree 
  object of class \code{bag}.
    The \code{grabmulti.dpart} function can only be used on 
    tree objects of class \code{dpart}.
      These functions can be invoked explicitly or just by 
      calling the \code{grab} 
        and \code{grabmulti} functions.
}
\examples{
# Assigning prey colours for default palette
  val <- apc(x = yftdiet, preyfile = PreyTaxonSort, check = TRUE)
  node.colsY <- val$cols
  dietPP <- val$x   # updated diet matrix with Group assigned prey taxa codes
  
# Fitting the classification tree
  yft.dp <- dpart(Group ~ Lat + Lon + Year + Quarter + SST  + Length, 
                      data = dietPP, weights = W, minsplit = 10,
                                        cp = 0.001)
                                       
# Pruning the tree
yft.pr <- prune(yft.dp, se = 1)                   

# Exploring Nodes: This suite of graphics is interactive and therefore has
# been commented out. When run, the code will ask you to select a node for
# viewing.
\dontrun{
  # Exploring nodes of the tree - single page
  val <- grab(object = yft.pr, LatID = "Lat", LonID = "Lon", setID = "TripSetNo", 
              node.cols = node.colsY, cex = 1, mapxlim = c(-125, -75), mapylim = c(0, 30),
                          mapcol = "gold3", pos = "topleft")
                          
 # Exploring nodes of the tree - multiple pages
 val <- grab(object = yft.pr, LatID = "Lat", LonID = "Lon", setID = "TripSetNo", 
             node.cols = node.colsY, cex = 1, mapxlim = c(-125, -75), mapylim = c(0, 30),
                         mapcol = "gold3", pos = "topleft", onepage = FALSE)
                         
 # Exploring multiple nodes
 grabmulti(object = yft.pr, LatID = "Lat", LonID = "Lon", setID = "TripSetNo", 
           node.cols = node.colsY, cex = 0.8, mapxlim = c(-125, -75), mapylim = c(0, 30))
}
}
\references{
Kuhnert, P.M., Duffy, L. M and Olson, R.J. (2012) The Analysis of Predator Diet 
and Stable Isotope Data, Journal of Statistical Software, In Prep.
}
\seealso{
\code{\link{plot.dpart}}; \code{\link{bagging}}
}
