% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\alias{importance.dpart}
\title{Variable Importance Ranking}
\usage{
\method{importance}{dpart}(x)

\method{importance}{dpart}(x)
}
\arguments{
\item{x}{object of class \code{dpart}}
}
\description{
Produces a variable importance ranking from a fitted \code{dpart} object.
}
\details{
A barplot of the variable importance ranking sorted from highest to lowest is produced 
in addition to the printed output.

Variable importance ranking for the variables used in the classification tree.
}
\examples{

# Assigning prey colours for default palette
val <- apc(x = yftdiet, preyfile = PreyTaxonSort, check = TRUE)
node.colsY <- val$cols
dietPP <- val$x   # updated diet matrix with Group assigned prey taxa codes

# Fitting the classification tree
yft.dp <- dpart(Group ~ Lat + Lon + Year + Quarter + SST  + Length, 
                 data = dietPP, weights = W, minsplit = 10,
                                    cp = 0.001)
yft.pr <- prune(yft.dp, se = 1)
vi <- importance(yft.pr)
}
\references{
Breiman, L., Friedman, J.H., Olshen, R.A. and Stone, C.J. (1984) Classification 
and Regression Trees. Wadsworth International.
}
\seealso{
\code{\link{dpart}}
}
