% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappoints.r
\name{mappoints}
\alias{mappoints}
\alias{mappoints.data}
\title{Produce a Map of a Region with Data overlayed}
\usage{
mappoints(XY.data, xlim, ylim, database = 'world', gtitle)

\method{mappoints}{data}(XY.data, xlim, ylim, database = "world",
  gtitle = NULL)
}
\arguments{
\item{XY.data}{matrix of longitude and latitude co-ordinates}

\item{xlim}{map x limits}

\item{ylim}{map y limits}

\item{database}{either "world" or "world2" depending on the area being plotted. Defaults to "world"}

\item{gtitle}{Title of Plot}
}
\description{
Utilises the \code{map} function to produce a map of the study area based on
locations given.
}
\details{
Calls on the \code{map} function for plotting.
}
\examples{
# Load dataset
data(yftdiet)

# Produce map of the data
mappoints(XY.data = yftdiet[,c("Lon", "Lat")])
#title(main = "YFT Data Distribution")

}
\references{
Richard A. Becker, and Allan R. Wilks, Maps in S, AT\&T Bell Laboratories 
Statistics Research Report [93.2], 1993. 

Richard A. Becker, and Allan R. Wilks, Constructing a Geographical Database, 
AT&T Bell Laboratories Statistics Research Report [95.2], 1995.
}
