% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPredator.r
\name{MapPredator}
\alias{MapPredator}
\title{MapPredator}
\usage{
MapPredator(x, LonID, LatID, mapxlim, mapylim, database, PredSpID)
}
\arguments{
\item{x}{x}

\item{LonID}{LonID}

\item{LatID}{LatID}

\item{mapxlim}{mapxlim}

\item{mapylim}{mapylim}

\item{database}{database}

\item{PredSpID}{PredSpID}
}
\description{
Internal function used by \code{plot.diet}
}
