% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPrey.r
\name{MapPrey}
\alias{MapPrey}
\title{MapPrey}
\usage{
MapPrey(x, LonID, LatID, mapxlim, mapylim, database)
}
\arguments{
\item{x}{x}

\item{LonID}{LonID}

\item{LatID}{LatID}

\item{mapxlim}{mapxlim}

\item{mapylim}{mapylim}

\item{database}{database}
}
\description{
Internal function used by \code{plot.diet}
}
