% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPreyPredator.r
\name{MapPreyPredator}
\alias{MapPreyPredator}
\title{MapPreyPredator}
\usage{
MapPreyPredator(x, PredSpID, LonID, LatID, mapxlim, mapylim, database)
}
\arguments{
\item{x}{x}

\item{PredSpID}{PredSpID}

\item{LonID}{LonID}

\item{LatID}{LatID}

\item{mapxlim}{mapxlim}

\item{mapylim}{mapylim}

\item{database}{database}
}
\description{
Internal function used in \code{plot.diet}
}
