% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partdep.r
\name{partdep}
\alias{partdep}
\alias{partdep.bag}
\title{Partial Dependence Plots}
\usage{
partdep(object, Xvar, Yvar = NULL, fact = FALSE, var.cond = NULL,
  plotmap = FALSE, mfrow = NULL, xlim = NULL, ylim = NULL,
  leg.pos = "topleft", plot2file = FALSE, se.fit = FALSE,
  too.far = 0.1, sp.id = NULL)

\method{partdep}{bag}(object, Xvar, Yvar = NULL, fact = FALSE,
  var.cond = NULL, plotmap = FALSE, mfrow = NULL, xlim = NULL,
  ylim = NULL, leg.pos = "topleft", plot2file = FALSE,
  se.fit = FALSE, too.far = 0.1, sp.id = NULL, palette)
}
\arguments{
\item{object}{object of class \code{bag}}

\item{Xvar}{name of variable/s to plot. This can be a single variable e.g. "Len" (1D plot) or
the name of two variables e.g. c("Lon", "Lat") for a 2D plot. Currenly only 2D plots
for longitude and latitude are implemented.}

\item{Yvar}{vector of prey names where partial dependence plots are required. If NULL (default)
then plots are produced for all prey.}

\item{fact}{logical. Is the variable a factor? This only works for 1D plots}

\item{var.cond}{name of the conditioning variable used to condition each plot by.}

\item{plotmap}{logical. Should a map be plotted? Only useful if longitude and latitude are being
plotted as a 2D plot. (default = FALSE)}

\item{mfrow}{plotting argument for 1D plots indicating the number of frames per row to plot 
(default: c(2,2)). See \code{\link{par}} for more details.}

\item{xlim}{range of values the x-axis takes when plotting. Defaults to the range of the x-data.
See \code{\link{par}} for more details.}

\item{ylim}{range of values the y-axis takes when plotting. Defaults to the range of the y-data.
See \code{\link{par}} for more details.}

\item{leg.pos}{position for the prey names in the 2D plots. Defaults to "topleft".}

\item{plot2file}{logical. Should the plots be written to file. If so, the file name defaults to "partdep.pdf"
in the current working directory.}

\item{se.fit}{logical. Should standard errors be produced and plotted on the figures of the 1D plots?
(default: FALSE)}

\item{too.far}{How far to clip the map (Default = 0.1)}

\item{sp.id}{Species ID (Default = NULL)}

\item{palette}{colour palette (created using the apc function)}
}
\value{
1D and 2D plots of partial dependence.
}
\description{
Produce partial dependence plots for examing each variable's contribution to the predicted response.
}
\details{
There are many different combinations of arguments for producing 1D and 2D plots.
Illustrations of the combinations are shown in the examples section below.
}
\examples{
# Assigning prey colours for default palette
#val <- apc(x = yftdiet, preyfile = PreyTaxonSort, check = TRUE)
#node.colsY <- val$cols
#dietPP <- val$x   # updated diet matrix with Group assigned prey taxa codes

# Fitting the classification tree
#yft.dp <- dpart(Group ~ Lat + Lon + Year + Quarter + SST  + Length, 
#data = dietPP, weights = W, minsplit = 10, cp = 0.001)
#yft.pr <- prune(yft.dp, se = 1)
#plot(yft.pr, node.cols = node.colsY)

# Bagging
# Bagging with NO spatial bootstrapping
#yft.bag <- bagging(Group ~ Lat + Lon + Year + Quarter + SST  + Length, data = dietPP, weights = W, minsplit = 50,
#             cp = 0.001, nBaggs = 500, predID = "TripSetPredNo")
            
# 1D plots based on covariates in tree model
#partdep(object = yft.bag, Xvar = "Length")
#partdep(object = yft.bag, Xvar = "Length", se.fit = TRUE)
#partdep(object = yft.bag, Xvar = "SST")
#partdep(object = yft.bag, Xvar = "Quarter", fact = TRUE, se.fit = TRUE)
#partdep(object = yft.bag, Xvar = "Year", fact = TRUE)

# 2D plots of Longitude and Latitude
#partdep(object = yft.bag, Xvar = c("Lon", "Lat"), plotmap = TRUE)
# 2D plots of Longitude and Latitude conditioning on Year
#partdep(object = yft.bag, Xvar = c("Lon", "Lat"), plotmap = TRUE, leg.pos="topleft", too.far = 0.05, sp.id = "F.Ost")
#partdep(object = yft.bag, Xvar = c("Lon", "Lat"), var.cond = list(Year = 2004),
#     too.far = 0.05, plotmap = TRUE, sp.id = "F.Ost")

# 2D plots of SST and Length
#partdep(object = yft.bag, Xvar = c("SST", "Length"))


}
\references{
Kuhnert, P.M., Duffy, L. M and Olson, R.J. (2012) The Analysis of Predator Diet 
and Stable Isotope Data, Journal of Statistical Software, In Prep. 

Kuhnert PM, Kinsey-Henderson A, Bartley R, Herr A (2010)
Incorporating uncertainty in gully erosion calculations using
the random forests modelling approach. Environmetrics
21:493-509. doi:10.1002/env.999
}
