% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpred.r
\name{readpred}
\alias{readpred}
\title{Internal function used by \code{read.pp} and \code{read.dm}}
\usage{
readpred(dat, diet.ind.start, sp2omit, FullnessL, DateL, Datef, p)
}
\arguments{
\item{dat}{A compositional data matrix containing covariate 
information and diet composition for each predator}

\item{diet.ind.start}{index where diet compositional data starts}

\item{sp2omit}{species to omit}

\item{FullnessL}{Fullness column name used to identify full stomachs for analysis}

\item{DateL}{Date column name}

\item{Datef}{date format}

\item{p}{Proportion of prey to omit.}
}
\description{
Internal function that extracts the species information
}
