% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart_internalFNs.r
\name{rpart.class}
\alias{rpart.class}
\alias{rpart.anova}
\alias{rpart.poisson}
\alias{rpartco}
\alias{plot.rpart}
\title{RPART functions}
\usage{
rpart.class(y, offset, parms, wt)

rpart.anova(y, offset, parms, wt)

rpart.poisson(y, offset, parms, wt)

rpartco(tree, parms)

\method{plot}{rpart}(x, uniform = FALSE, branch = 1,
  compress = FALSE, nspace, margin = 0, minbranch = 0.3, ...)
}
\description{
These have been exported from the rpart package to
avoid errors in the check
}
