% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmPlots.r
\name{SmPlots}
\alias{SmPlots}
\title{Smooth Plots}
\usage{
SmPlots(x, i, SmXvar, SmXdat, LonID, LatID, projection, palette,
  too.far = 0.05)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{SmXvar}{SmXvar}

\item{SmXdat}{SmXdat}

\item{LonID}{LonID}

\item{LatID}{LatID}

\item{projection}{projection}

\item{palette}{palette}

\item{too.far}{too.far}
}
\description{
Smooth plots used in \code{plot.diet}
}
