\name{spatialsamp}
\alias{spatialsamp}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
spatial samp
}

\description{
concise (1-5 lines) description of what the function does.
}
\usage{
spatialsamp(x, LonID, LatID, sizeofgrid = 5, ID = NULL, nsub = NULL, Plot = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Describe \code{x} here
}
\item{LonID}{
Describe \code{x} here
}
\item{LatID}{
Describe \code{x} here
}
  \item{sizeofgrid}{
Describe \code{x} here
}
  \item{ID}{
Describe \code{x} here
}
  \item{nsub}{
Describe \code{x} here
}
  \item{Plot}{
Describe \code{x} here
}
 
}
\details{
If necessary, more details than the description above
}
\value{
Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
put references to the literature/web site here ~
}
\author{
who you are
}
\note{
further notes
}


\seealso{
objects to See Also as \code{\link{help}},
}
\examples{
ls()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
