\name{subsample}
\alias{subsample}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
subsample
}

\description{
concise (1-5 lines) description of what the function does.
}
\usage{
subsample(dat, ID, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
Describe \code{x} here
}
  \item{ID}{
Describe \code{x} here
}
  \item{n}{
Describe \code{\dots} here
}
}
\details{
If necessary, more details than the description above
}
\value{
Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
put references to the literature/web site here ~
}
\author{
who you are
}
\note{
further notes
}


\seealso{
objects to See Also as \code{\link{help}},
}
\examples{
ls()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
