% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dpart.R
\name{summary}
\alias{summary}
\alias{summary.dpart}
\alias{summary.rpart}
\title{Summarise a Fitted dpart Object}
\usage{
summary(object, ...)

\method{summary}{dpart}(object, cp = 0, digits = getOption("digits"),
  file = "tree_output.txt", ...)

\method{summary}{rpart}(object, cp = 0, digits = getOption("digits"),
  file, ...)
}
\arguments{
\item{object}{dpart object}

\item{...}{additional arguments affecting the summary produced}

\item{cp}{trim nodes with a complexity of less than cp from the listing.}

\item{digits}{Number of significant digits to be used in the result.}

\item{file}{write the output to a given file name. (Full listings of a tree are often quite long).
Default file name is \code{tree_output.txt}.}
}
\description{
Outputs a detailed listing of the classification tree to file: tree_output.txt
}
\details{
This function is a method for the generic function summary for class \code{dpart}. 
It can be invoked by calling summary for an object of the appropriate class, 
or directly by calling \code{summary.dpart} regardless of the class of the object. 
By default the summary is written to file.
}
\examples{

# Assigning prey colours for default palette
val <- apc(x = yftdiet, preyfile = PreyTaxonSort, check = TRUE)
node.colsY <- val$cols
dietPP <- val$x   # updated diet matrix with Group assigned prey taxa codes

# Fitting the classification tree
yft.dp <- dpart(Group ~ Lat + Lon + Year + Quarter + SST  + Length, 
                  data = dietPP, weights = W, minsplit = 10,
                                    cp = 0.001)
                                    
plot(yft.dp, node.cols = node.colsY)
summary(yft.dp)
#print(yft.dp, setID = "TripSetNo")
}
\seealso{
\code{\link{summary}}, \code{\link{dpart}}, \code{\link{print.dpart}}
}
