% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcountries.r
\docType{data}
\name{worldcountries}
\alias{worldcountries}
\title{worldcountries}
\format{An object of class \code{SpatialPolygonsDataFrame} with 177 rows and 62 columns.}
\usage{
worldcountries
}
\description{
@description 
 #'  A SpatialPolygonsDataFrame containing world country information
 that was exported from the \code{spaMM} package.
}
\keyword{datasets}
