% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet-Internal.r
\name{rpart.matrix}
\alias{rpart.matrix}
\alias{rpart.branch}
\alias{rpconvert}
\alias{SpeciesCompBar}
\alias{subsample}
\alias{spatialsamp}
\alias{formOmat}
\alias{Vis.Gam}
\alias{vis}
\alias{vis.igam}
\alias{textG.dpart}
\alias{text.dpart}
\alias{tree.depth}
\alias{tree.depth.dpart}
\alias{snip.dpart}
\alias{snip.dpart.mouse}
\alias{select.tree.dpart}
\alias{select.tree}
\alias{rsq.dpart}
\alias{rsq}
\alias{rpartco.dpart}
\alias{plotG.dpart}
\alias{outside}
\alias{na.igam}
\alias{na.dpart}
\alias{mask}
\alias{explore.bag}
\alias{explore}
\alias{writepn.csv}
\alias{Distance}
\title{Internal functions used in the diet package}
\usage{
rpart.matrix(frame)

rpart.branch(x, y, node, branch)

rpconvert(x)

SpeciesCompBar(x, prey.cols, Factor, Species)

subsample(dat, ID, n)

spatialsamp(x, LonID, LatID, sizeofgrid = 5, ID = NULL, nsub = NULL,
  Plot = FALSE)

formOmat(object, ID)

Vis.Gam(x, view = NULL, cond = list(), n.grid = 30, too.far = 0,
  col = NA, color = "heat", contour.col = NULL, se = -1,
  type = "link", plot.type = "persp", zlim = NULL, nCol = 50, ...)

vis(x, view = NULL, cond = list(), n.grid = 30, too.far = 0,
  col = NA, color = "heat", contour.col = NULL, se = -1,
  type = "link", plot.type = "persp", zlim = NULL, nCol = 50,
  se.plot = TRUE, ...)

\method{vis}{igam}(x, view = NULL, cond = list(), n.grid = 30,
  too.far = 0, col = NA, color = "heat", contour.col = NULL,
  se = -1, type = "link", plot.type = "persp", zlim = NULL,
  nCol = 50, se.plot = TRUE, ...)

textG.dpart(x, splits = TRUE, which = 4, label = "yval",
  FUN = text, all.leaves = FALSE, pretty = NULL,
  digits = getOption("digits") - 2, tadj = 0.65, use.n = FALSE,
  bars = TRUE, xadj = 1, yadj = 1, bord = FALSE,
  node.cols = NULL, pos = NULL, ...)

\method{text}{dpart}(x, splits = TRUE, which = 4, label = "yval",
  FUN = text, all.leaves = FALSE, pretty = NULL,
  digits = getOption("digits") - 2, tadj = 0.65, use.n = FALSE,
  bars = TRUE, xadj = 1, yadj = 1, bord = FALSE,
  node.cols = NULL, pos = NULL, ...)

tree.depth(nodes)

tree.depth.dpart(nodes)

snip.dpart(x, toss)

snip.dpart.mouse(tree, parms = paste(".rpart.parms", dev.cur(), sep =
  "."))

\method{select.tree}{dpart}(object, se, nsplits)

select.tree(object, ...)

\method{rsq}{dpart}(x, rn)

rsq(x, ...)

rpartco.dpart(tree, parms = paste(".rpart.parms", dev.cur(), sep = "."))

plotG.dpart(x, node.cols = NULL, pos = NULL, ...)

outside(polyx, polyy, x, y)

na.igam(x)

na.dpart(x)

mask(gridx, gridy, x, y)

explore.bag(object, node, cols = NULL, showtitle = FALSE,
  axis.side = 2, cex = 1, ylim)

explore(object, pred, pred.where, loss = NULL, node, cols = NULL,
  showtitle = FALSE, labels = TRUE, cex = 1, ylim)

writepn.csv(x)

Distance(O, P, type = "Hellinger")
}
\description{
These are internal functions used in the \code{diet} packagethat are not intended to 
be directly called by the user.
}
\keyword{internal}
